# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Education Planner API",
    "description": "\n    ## Динамический планировщик образовательной траектории\n    \n    API для студентов УрФУ, помогающее:\n    - 📚 Выбирать специальность и курсы\n    - 📊 Отслеживать развитие навыков\n    - 🎯 Получать персональные рекомендации\n    - 🔄 Моделировать смену специальности\n    \n    ### Основные возможности:\n    \n    - **Аутентификация**: JWT токены (access + refresh)\n    - **Специальности**: Просмотр roadmap и матрицы навыков\n    - **Курсы**: Запись, прохождение, отслеживание прогресса\n    - **Навыки**: Древовидная структура, gap-анализ\n    - **Рекомендации**: Умные подсказки по выбору курсов\n    - **What-If анализ**: Моделирование смены специальности\n    \n    ### Для начала работы:\n    1. Зарегистрируйтесь через `/api/auth/register`\n    2. Получите токен через `/api/auth/login`\n    3. Используйте токен в заголовке: `Authorization: Bearer \u003Ctoken\u003E`\n    ",
    "contact": {
      "name": "Education Planner Team",
      "email": "support@edu-planner.ru"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/specializations/": {
      "get": {
        "tags": [
          "specializations"
        ],
        "summary": "Get All Specializations",
        "description": "Получить список всех специальностей",
        "operationId": "get_all_specializations_api_specializations__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SpecializationResponse"
                  },
                  "type": "array",
                  "title": "Response Get All Specializations Api Specializations  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "specializations"
        ],
        "summary": "Create Specialization",
        "description": "Создать новую специальность (для администраторов)",
        "operationId": "create_specialization_api_specializations__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecializationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecializationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/specializations/{specialization_id}": {
      "get": {
        "tags": [
          "specializations"
        ],
        "summary": "Get Specialization",
        "description": "Получить детальную информацию о специальности",
        "operationId": "get_specialization_api_specializations__specialization_id__get",
        "parameters": [
          {
            "name": "specialization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Specialization Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecializationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/specializations/{specialization_id}/roadmap": {
      "get": {
        "tags": [
          "specializations"
        ],
        "summary": "Get Specialization Roadmap",
        "description": "Получить roadmap специальности на весь период обучения",
        "operationId": "get_specialization_roadmap_api_specializations__specialization_id__roadmap_get",
        "parameters": [
          {
            "name": "specialization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Specialization Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoadmapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/specializations/{specialization_id}/skills-matrix": {
      "get": {
        "tags": [
          "specializations"
        ],
        "summary": "Get Skills Matrix",
        "description": "Получить матрицу требуемых навыков для специальности",
        "operationId": "get_skills_matrix_api_specializations__specialization_id__skills_matrix_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "specialization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Specialization Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillsMatrixResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/courses/{course_id}/skills": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get Course Skills",
        "description": "Получить навыки курса по уровням сложности",
        "operationId": "get_course_skills_api_courses__course_id__skills_get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/skills/my-skills": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "Get My Skills",
        "description": "Получить текущие навыки студента",
        "operationId": "get_my_skills_api_skills_my_skills_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StudentSkillResponse"
                  },
                  "type": "array",
                  "title": "Response Get My Skills Api Skills My Skills Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/skills/gap-analysis": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "Get Gap Analysis",
        "description": "Анализ дефицита навыков для выбранной специальности",
        "operationId": "get_gap_analysis_api_skills_gap_analysis_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "specialization_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID специальности для анализа (если не указан, используется специальность студента)",
              "title": "Specialization Id"
            },
            "description": "ID специальности для анализа (если не указан, используется специальность студента)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GapAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/students/me/specialization": {
      "put": {
        "tags": [
          "students"
        ],
        "summary": "Set Specialization",
        "description": "Выбрать или сменить специальность",
        "operationId": "set_specialization_api_students_me_specialization_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "specialization_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Specialization Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/students/me/resume": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "Get My Resume",
        "description": "Получить резюме студента на основе изученных курсов и навыков",
        "operationId": "get_my_resume_api_students_me_resume_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/resume/": {
      "get": {
        "tags": [
          "resume"
        ],
        "summary": "Get My Resume",
        "description": "Получить текущее резюме студента.\n\nРезюме включает:\n- личную информацию (имя, email, основная специальность)\n- все навыки студента из:\n  - основной специальности (core)\n  - всех майнеров, включая архивные (orbit)\n- информацию о майнерах (текущие и завершённые)\n- статистику образования (курсы, кредиты)\n- достижения\n\n**Важно**: даже если майнер был изменён (archived),\nвсе навыки из него остаются в резюме!\n\nЭто резюме можно сгенерировать в любой момент обучения\nили по окончании основной специальности как итоговое.",
        "operationId": "get_my_resume_api_resume__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CourseDetailResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "credits": {
            "type": "integer",
            "maximum": 10,
            "minimum": 1,
            "title": "Credits",
            "default": 3
          },
          "semester": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 12,
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Semester"
          },
          "is_elective": {
            "type": "boolean",
            "title": "Is Elective",
            "default": false
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "specialization_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialization Id"
          },
          "difficulty_levels": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty Levels"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillInCourse"
            },
            "type": "array",
            "title": "Skills",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "CourseDetailResponse",
        "description": "Детальная информация о курсе"
      },
...
      "CourseInRoadmap": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "semester": {
            "type": "integer",
            "title": "Semester"
          },
          "is_elective": {
            "type": "boolean",
            "title": "Is Elective"
          },
          "credits": {
            "type": "integer",
            "title": "Credits"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "semester",
          "is_elective",
          "credits"
        ],
        "title": "CourseInRoadmap",
        "description": "Курс в roadmap"
      },
...
      "GapAnalysisItem": {
        "properties": {
          "skill_id": {
            "type": "integer",
            "title": "Skill Id"
          },
          "skill_name": {
            "type": "string",
            "title": "Skill Name"
          },
          "required_level": {
            "type": "integer",
            "title": "Required Level"
          },
          "current_level": {
            "type": "integer",
            "title": "Current Level"
          },
          "gap": {
            "type": "integer",
            "title": "Gap"
          },
          "is_satisfied": {
            "type": "boolean",
            "title": "Is Satisfied"
          }
        },
        "type": "object",
        "required": [
          "skill_id",
          "skill_name",
          "required_level",
          "current_level",
          "gap",
          "is_satisfied"
        ],
        "title": "GapAnalysisItem",
        "description": "Элемент анализа дефицита навыков"
      },
      "GapAnalysisResponse": {
        "properties": {
          "specialization_id": {
            "type": "integer",
            "title": "Specialization Id"
          },
          "specialization_name": {
            "type": "string",
            "title": "Specialization Name"
          },
          "total_required_skills": {
            "type": "integer",
            "title": "Total Required Skills"
          },
          "satisfied_skills": {
            "type": "integer",
            "title": "Satisfied Skills"
          },
          "completion_percentage": {
            "type": "number",
            "title": "Completion Percentage"
          },
          "gaps": {
            "items": {
              "$ref": "#/components/schemas/GapAnalysisItem"
            },
            "type": "array",
            "title": "Gaps"
          }
        },
        "type": "object",
        "required": [
          "specialization_id",
          "specialization_name",
          "total_required_skills",
          "satisfied_skills",
          "completion_percentage",
          "gaps"
        ],
        "title": "GapAnalysisResponse",
        "description": "Результат анализа дефицита навыков"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResumeResponse": {
        "properties": {
          "personal": {
            "$ref": "#/components/schemas/ResumePersonalInfo"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/ResumeSkill"
            },
            "type": "array",
            "title": "Skills"
          },
          "minors": {
            "items": {
              "$ref": "#/components/schemas/ResumeMinor"
            },
            "type": "array",
            "title": "Minors",
            "default": []
          },
          "education": {
            "$ref": "#/components/schemas/ResumeEducation"
          },
          "achievements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Achievements"
          }
        },
        "type": "object",
        "required": [
          "personal",
          "skills",
          "education",
          "achievements"
        ],
        "title": "ResumeResponse",
        "description": "Полное резюме студента (основа + орбита)"
      },
...
      "RoadmapResponse": {
        "properties": {
          "specialization_id": {
            "type": "integer",
            "title": "Specialization Id"
          },
          "specialization_name": {
            "type": "string",
            "title": "Specialization Name"
          },
          "duration_years": {
            "type": "integer",
            "title": "Duration Years"
          },
          "semesters": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/CourseInRoadmap"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Semesters"
          },
          "total_credits": {
            "type": "integer",
            "title": "Total Credits"
          }
        },
        "type": "object",
        "required": [
          "specialization_id",
          "specialization_name",
          "duration_years",
          "semesters",
          "total_credits"
        ],
        "title": "RoadmapResponse",
        "description": "Roadmap образовательной траектории"
      },
      "SkillInCourse": {
        "properties": {
          "skill_id": {
            "type": "integer",
            "title": "Skill Id"
          },
          "skill_name": {
            "type": "string",
            "title": "Skill Name"
          },
          "difficulty_level": {
            "type": "integer",
            "title": "Difficulty Level"
          },
          "skill_weight": {
            "type": "integer",
            "title": "Skill Weight"
          }
        },
        "type": "object",
        "required": [
          "skill_id",
          "skill_name",
          "difficulty_level",
          "skill_weight"
        ],
        "title": "SkillInCourse",
        "description": "Навык в курсе"
      },
      "SkillRequirement": {
        "properties": {
          "skill_id": {
            "type": "integer",
            "title": "Skill Id"
          },
          "skill_name": {
            "type": "string",
            "title": "Skill Name"
          },
          "required_level": {
            "type": "integer",
            "title": "Required Level"
          },
          "current_level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Level",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "skill_id",
          "skill_name",
          "required_level"
        ],
        "title": "SkillRequirement",
        "description": "Требуемый навык"
      },
...
      "SkillsMatrixResponse": {
        "properties": {
          "specialization_id": {
            "type": "integer",
            "title": "Specialization Id"
          },
          "specialization_name": {
            "type": "string",
            "title": "Specialization Name"
          },
          "required_skills": {
            "items": {
              "$ref": "#/components/schemas/SkillRequirement"
            },
            "type": "array",
            "title": "Required Skills"
          }
        },
        "type": "object",
        "required": [
          "specialization_id",
          "specialization_name",
          "required_skills"
        ],
        "title": "SkillsMatrixResponse",
        "description": "Матрица навыков специальности"
      },
      "SpecializationCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "market_demand": {
            "type": "number",
            "maximum": 100,
            "minimum": 0,
            "title": "Market Demand",
            "default": 0
          },
          "duration_years": {
            "type": "integer",
            "maximum": 6,
            "minimum": 1,
            "title": "Duration Years",
            "default": 2
          },
          "required_skills": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Skills"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SpecializationCreate",
        "description": "Схема для создания специальности"
      },
      "SpecializationResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "market_demand": {
            "type": "number",
            "maximum": 100,
            "minimum": 0,
            "title": "Market Demand",
            "default": 0
          },
          "duration_years": {
            "type": "integer",
            "maximum": 6,
            "minimum": 1,
            "title": "Duration Years",
            "default": 2
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "required_skills": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Skills"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "SpecializationResponse",
        "description": "Схема ответа со специальностью"
      },
...
      "StudentSkillResponse": {
        "properties": {
          "skill_id": {
            "type": "integer",
            "title": "Skill Id"
          },
          "skill_name": {
            "type": "string",
            "title": "Skill Name"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "acquired_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acquired From"
          }
        },
        "type": "object",
        "required": [
          "skill_id",
          "skill_name",
          "level"
        ],
        "title": "StudentSkillResponse",
        "description": "Навык студента"
      },
...
      "UserResponse": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Full Name"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "specialization_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialization Id"
          },
          "minor_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Minor Ids"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "email",
          "full_name",
          "id",
          "is_active",
          "created_at"
        ],
        "title": "UserResponse",
        "description": "Схема ответа с данными пользователя"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}